[WIP] GPU support
Work in progress!

We now have something that seems to run all experiments (i.e. with Tensorflow and Pytorch after one another), but we need to make sure it also still runs on CPU. Also, it seems like only CLIP experiments are actually running on GPU, while the Tensorflow (imagenet) experiments seem not to. We are now running them from multiprocessing, so perhaps that doesn't jive with tensorflow?